fix(Tinebase/TempFile): only remove subdir if it is empty
see https://stackoverflow.com/a/18856880/8596553 for this solution

+++ tine20/Tinebase/TempFile.php
@@ -315,8 +315,9 @@ class Tinebase_TempFile extends Tinebase_Backend_Sql_Abstract implements Tinebas
      * @param string $dir
      * @param Tinebase_DateTime $date
      * @return int
+     * @throws Exception
      */
-    protected function _removeFilesFromDirByTimestamp($dir, Tinebase_DateTime $date)
+    protected function _removeFilesFromDirByTimestamp(string $dir, Tinebase_DateTime $date): int
     {
         if (Tinebase_Core::isLogLevel(Zend_Log::DEBUG)) Tinebase_Core::getLogger()->debug(__METHOD__ . '::' . __LINE__
             . ' Deleting old files from dir: ' . $dir);
@@ -341,11 +342,10 @@ class Tinebase_TempFile extends Tinebase_Backend_Sql_Abstract implements Tinebas
                 } catch (UnexpectedValueException $uve) {
                     // sub dir was already removed...
                 }
-                try {
+                $subDirIterator = new FilesystemIterator($directoryIterator->getPathname());
+                $isDirEmpty = !$subDirIterator->valid();
+                if ($isDirEmpty) {
                     rmdir($directoryIterator->getPathname());
-                } catch (Throwable $t) {
-                    if (Tinebase_Core::isLogLevel(Zend_Log::NOTICE)) Tinebase_Core::getLogger()->notice(
-                        __METHOD__ . '::' . __LINE__ . ' ' . $t->getMessage());
                 }
             }
